requirements.txt
appdirs==1.4.0
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
requests==2.13.0
django_rest_swagger==2.1.1

psycopg2

gunicorn
dj-static
